refactor auth btn
