+++ src/app/auth/auth-button/component.html
@@ -1,0 +1,22 @@
+<!-- log-out btn + 'gå til min side' -->
+<ng-container *ngIf="auth.isAuthenticated$ | async; else loggedOut">
+  <a class="lls-link" routerLink="/my-page">
+    Gå til min side
+  </a>
+  <a
+    (click)="auth.logout({ returnTo: document.location.origin })"
+    class="lls-link u-ml-4"
+  >
+    Log ud
+  </a>
+</ng-container>
+
+<!-- log-in btn -->
+<ng-template #loggedOut>
+  <a
+    (click)="login()"
+    class="lls-link"
+  >
+    Log ind
+  </a>
+</ng-template>
