+++ src/app/auth/auth-button/component.ts
@@ -4,29 +4,7 @@ import { DOCUMENT } from '@angular/common';
 
 @Component({
   selector: 'lls-auth-button',
-  template: `
-    <ng-container *ngIf="auth.isAuthenticated$ | async; else loggedOut">
-      <a class="lls-link" routerLink="/my-page">
-        Gå til min side
-      </a>
-      <a
-        (click)="auth.logout({ returnTo: document.location.origin })"
-        class="lls-link u-ml-4"
-      >
-        Log ud
-      </a>
-    </ng-container>
-
-    <ng-template #loggedOut>
-      <a
-        (click)="login()"
-        class="lls-link"
-      >
-        Log ind
-      </a>
-    </ng-template>
-  `,
-  styles: [],
+  templateUrl: './component.html',
 })
 
 export class AuthButtonComponent {
